Updates adapt_by_field to allow possibility of turing coarsening off
The `sizefieldParam` file can either have 13 or 14 inputs. Based on the
number of inputs the following behaviours are expected:

(a) if the number of inputs in the file is 13, the behaviour is the same
as before
(b) if the number of inputs in the file is 14, the first 13 will be used
as before, and coarsening will be turned off. (note that the value of
the 14th parameter has no effect, only the fact that it exists)